Extract content selection into separate component
Extract the logic for presenting a list of available content sources and
showing the dialogs for choosing content from specific sources into a
separate `ContentSelector` component. This will keep FilePickerApp more
manageable as new steps are added to the assignment configuration flow.

 - Extract logic, JSX and associated styles for content selection buttons into
   `ContentSelector` component. Also move the associated tests and
   `Config` context values from the `FilePickerApp` tests to the
   `ContentSelector` tests

 - Change the `lms` enum value of the `DialogType` enum to the more
   descriptive `lmsFile`

 - Rename the `data-testid` values of several of the content selection
   buttons to slightly more descriptive names
