+++ lms/static/styles/frontend_apps.scss
@@ -11,6 +11,7 @@
 
 // Preact components
 @use 'components/BasicLtiLaunchApp';
+@use 'components/ContentSelector';
 @use 'components/Dialog';
 @use 'components/ErrorDisplay';
 @use 'components/FileList';
